+++ src/babble/babble.go
@@ -1,5 +1,4 @@
 // Package babble wraps the babble and EVM-Lite components.
-
 package babble
 
 import (